+++ src/test/java/com/github/havarunner/sequentiality/SuiteMember.java
@@ -1,0 +1,12 @@
+package com.github.havarunner.sequentiality;
+
+import com.github.havarunner.annotation.PartOf;
+import org.junit.Test;
+
+@PartOf(SequentialSuite.class)
+class SuiteMember {
+    SuiteMember(String suiteObject) {}
+
+    @Test
+    void test() {}
+}